# strutil: return the condition instead of returning based on condition