Quick change where we return the condition of the if statement instead of checking then returning true or false.

NOTE: Added extra benchmark and fixed a lint 
```
name              old time/op    new time/op    delta
HasSuffixFold-10    26.6ns ± 1%    26.2ns ± 1%  -1.22%  (p=0.000 n=10+10)
HasPrefixFold-10    22.3ns ± 1%    21.7ns ± 0%  -2.45%  (p=0.000 n=8+8)
ContainsFold-10     21.1ns ± 0%    20.4ns ± 0%  -3.03%  (p=0.000 n=9+10)

name              old alloc/op   new alloc/op   delta
HasSuffixFold-10     0.00B          0.00B         ~     (all equal)
HasPrefixFold-10     0.00B          0.00B         ~     (all equal)
ContainsFold-10      0.00B          0.00B         ~     (all equal)

name              old allocs/op  new allocs/op  delta
HasSuffixFold-10      0.00           0.00         ~     (all equal)
HasPrefixFold-10      0.00           0.00         ~     (all equal)
ContainsFold-10       0.00           0.00         ~     (all equal)
```