@puellanivis 2023-09-20
What is the performances shown by `BenchmarkContainsFoldToLower`? (I imagine the intent is to compare benchmarks of it vs `ContainsFold` right?)