@seiflotfy 2023-09-20
The intent is to show that ContainsFold is now faster than before... ContainsFold is already faster than ContainsFoldToLower. Or am I missing something?