@puellanivis 2023-09-21
You’re not missing anything. I just hope to see what the metrics actually look like.